+++ wetterdienst/core/util.py
@@ -2,27 +2,24 @@
 # Distributed under the MIT License. See LICENSE for more info.
 """Tools for timeseries."""
 
-from __future__ import annotations
+from dataclasses import dataclass
 
 import polars as pl
 
 
+@dataclass
 class _ParameterData:
-    def __init__(
-        self,
-        values: pl.DataFrame,
-        station_ids: list[str] | None = None,
-        extra_station_counter: int = 0,
-    ) -> None:
-        self.station_ids = station_ids or []
-        self.extra_station_counter = extra_station_counter
-        self.values = values
-        self.finished = False
+    values: pl.DataFrame
+    station_ids: list[str] | None = None
+    additional_station_counter: int = 0
+    finished: bool = False
 
 
 def extract_station_values(
     param_data: _ParameterData,
     result_series_param: pl.Series,
+    min_gain_of_value_pairs: float,
+    num_additional_stations: int,
     *,
     valid_station_groups_exists: bool,
 ) -> None:
@@ -31,23 +28,18 @@ def extract_station_values(
     # 1. only add further stations if not a minimum of 4 stations is reached OR
     # 2. a gain of 10% of timestamps with at least 4 existing values over all stations is seen OR
     # 3. an additional stations_counter is below 3 (used if a station has really no or few values)
-    cond1 = param_data.values.shape[1] < 4
-    cond2 = not cond1 and gain_of_value_pairs(param_data.values, result_series_param) > 0.10
-    if (
-        not valid_station_groups_exists or cond1 or cond2 or param_data.extra_station_counter < 3
-    ):  # timestamps + 4 stations
+    cond1 = param_data.values.shape[1] < 5  # 5: dates plus 4 values
+    cond2 = calculate_gain_of_value_pairs(param_data.values, result_series_param) >= min_gain_of_value_pairs
+    cond3 = param_data.additional_station_counter < num_additional_stations
+    if not valid_station_groups_exists or cond1 or cond2 or cond3:  # timestamps + 4 stations
         if not (cond1 or cond2):
-            param_data.extra_station_counter += 1
-        # "S" is added to station id titles to prevent bug with pandas that somehow doesn't allow column name "02000"
-        # under certain circumstances
-        param_data.values = param_data.values.with_columns(
-            pl.lit(result_series_param).alias(f"S{result_series_param.name}"),
-        )
+            param_data.additional_station_counter += 1
+        param_data.values = param_data.values.with_columns(result_series_param)
     else:
         param_data.finished = True
 
 
-def gain_of_value_pairs(old_values: pl.DataFrame, new_values: pl.Series) -> float:
+def calculate_gain_of_value_pairs(old_values: pl.DataFrame, new_values: pl.Series) -> float:
     """Calculate the gain of value pairs.
 
     The gain of value pairs is calculated by the following formula:
@@ -57,13 +49,13 @@ def gain_of_value_pairs(old_values: pl.DataFrame, new_values: pl.Series) -> floa
 
     """
     old_score = (
-        old_values.select(pl.fold(acc=0, function=lambda acc, s: acc + s.is_not_null(), exprs=pl.all()) >= 4)
+        old_values.select(pl.fold(acc=0, function=lambda acc, s: acc + s.is_not_null(), exprs=pl.all()) >= 5)
         .sum()
         .item()
     )  # 5: dates plus 4 values
-    old_values = old_values.with_columns(pl.lit(new_values).alias(f"S{new_values.name}"))
+    old_values = old_values.with_columns(pl.lit(new_values).alias(new_values.name))
     new_score = (
-        old_values.select(pl.fold(acc=0, function=lambda acc, s: acc + s.is_not_null(), exprs=pl.all()) >= 4)
+        old_values.select(pl.fold(acc=0, function=lambda acc, s: acc + s.is_not_null(), exprs=pl.all()) >= 5)
         .sum()
         .item()
     )  # 5: dates plus 4 values